removed old createShowings.js
updated readme
