+++ README.md
@@ -2,4 +2,14 @@
 
 Begin with ```npm install```
 
-First starts mongoDB, next starts server (index.js)
+Continue with following steps:
+1. starts mongoDB
+2. execute ```node importer```
+3. execute ```node addPricesToDB```
+4. execute ```node addSaloonsToDB```
+5. execute ```node updateSaloonsBestRows```
+6. execute ```node createShowings```
+7. start application with ```nodemon``` or ```node index```
+
+
+
