+++ createShowings.js
@@ -14,8 +14,11 @@ let Saloon = require('./models/Saloon');
 let Film = require('./models/Film');
 let Showing = require('./models/Showing');
 
+function randomItem(arr) {
+  return arr[Math.floor(Math.random() * arr.length)];
+}
 
-let shuffleArr = (array) => {
+function shuffleArr(array){
   for (let i = array.length - 1; i > 0; i--) {
     let j = Math.floor(Math.random() * (i + 1));
     let temp = array[i];
@@ -24,11 +27,6 @@ let shuffleArr = (array) => {
   }
   return array;
 }
-
-function randomItem(arr){
-  return arr[Math.floor(Math.random()*arr.length)];
-}
-
 async function createAndAddShowingsTo() {
   let showingsCount = await Showing.count();
 
@@ -40,24 +38,69 @@ async function createAndAddShowingsTo() {
 
   let movies = await Film.find();
   let saloons = await Saloon.find();
-
+  let times = ['17:00', '19:30', '22:00'];
   for (let i = 0; i < 84; i++) {
-
     if (i % 3 == 0) {
-      date.setDate(date.getDate() + 1);
+      date = new Date(date.setDate(date.getDate() + 1));
+      date.setHours(00);
+      date.setMinutes(00);
+      date.setSeconds(00);
     }
 
-    let film = randomItem(movies);
-    let saloon = randomItem(saloons);
-
     let showtime = new Showing({
-      "saloon": saloon,
-      "film": film,
       "date": date.getTime(),
-      "time": 17 + Math.floor(Math.random() * 4) + ':' + (Math.round(Math.random() < 0.5 ? 30 : 45))
     });
     await showtime.save();
+
   }
+
+  let showings = await Showing.find();
+  for (let i = 0; i < showings.length; i++) {
+    let showing = showings[i];
+    let moviesShuffle = shuffleArr(movies);
+    let film = randomItem(moviesShuffle);
+    showing.film = film;
+    let saloon;
+    let time;
+    if (i % 2 === 0) {
+      saloon = saloons[0];
+      time = times[1];
+      showing.saloon = saloon;
+      showing.time = time;
+      moviesShuffle = shuffleArr(movies);
+
+      film = randomItem(moviesShuffle);
+      showing.film = film;
+    }
+
+
+    else {
+    saloon = saloons[2];
+    time = times[0];
+    showing.saloon = saloon;
+    showing.time = time;
+    moviesShuffle = shuffleArr(movies);
+
+    film = randomItem(moviesShuffle);
+    showing.film = film;
+    }
+    await showing.save();  
+  }
+
+  showings = await Showing.find();
+  for (let i = 0; i < showings.length; i+=3) {
+    let showing = showings[i];
+
+    let saloon = saloons[1];
+    let time = times[2];
+    showing.saloon = saloon;
+    showing.time = time;
+    let film = randomItem(movies);
+    showing.film = film;
+
+    await showing.save();
+  }
+
 
   showingsCount = await Showing.count();
